generalized sidebar component
